fix: cast SUM(bigint) to bigint in reasoning repo and fix test ordering
SUM(bigint) returns numeric in PostgreSQL, which has variable-length
encoding that can exceed 8 bytes and fail sqlx i64 decoding. Add
explicit ::bigint casts to COALESCE(SUM(...), 0) in get_daily_spend_by_task
and get_spend_summary.

Also fix get_team_contributions_for_period test: insert contributions
BEFORE assigning person to team so assign_person_to_team picks up the
historical created_at date for membership start_date.

+++ crates/ps-core/src/repo/reasoning.rs
@@ -278,8 +278,8 @@ impl ReasoningRepo {
             SELECT
                 task_type,
                 COALESCE(SUM(estimated_cost_usd::double precision), 0.0) as "total_cost_usd!: f64",
-                COALESCE(SUM(prompt_tokens::bigint), 0) as "total_prompt_tokens!: i64",
-                COALESCE(SUM(completion_tokens::bigint), 0) as "total_completion_tokens!: i64",
+                COALESCE(SUM(prompt_tokens::bigint), 0)::bigint as "total_prompt_tokens!: i64",
+                COALESCE(SUM(completion_tokens::bigint), 0)::bigint as "total_completion_tokens!: i64",
                 COUNT(*) as "request_count!: i64"
             FROM reasoning.api_usage
             WHERE created_at::date = $1
@@ -317,8 +317,8 @@ impl ReasoningRepo {
                 model,
                 task_type,
                 COALESCE(SUM(estimated_cost_usd::double precision), 0.0) as "total_cost_usd!: f64",
-                COALESCE(SUM(prompt_tokens::bigint), 0) as "total_prompt_tokens!: i64",
-                COALESCE(SUM(completion_tokens::bigint), 0) as "total_completion_tokens!: i64",
+                COALESCE(SUM(prompt_tokens::bigint), 0)::bigint as "total_prompt_tokens!: i64",
+                COALESCE(SUM(completion_tokens::bigint), 0)::bigint as "total_completion_tokens!: i64",
                 COUNT(*) as "request_count!: i64"
             FROM reasoning.api_usage
             WHERE created_at >= $1 AND created_at < $2
